Exception Handling
In python while compiling if compiler found logical error then program is terminate from that line, this all handle by Python Data management. In python there is pre defined classes which is designed for handling different type of exception or error come while running the code.

- so if python handle the exception then it terminate the whole program if face any logical error.
Solution: is user handle the exception in spite of python.

- How user can handle the exception?
* by using some keywords:
- try
-except
-handle
- raise

below are the some examples how we use this keyword for exceptional handling in python.